Added Connection trait for general connection
The connection trait will be implemented for different connection types
(Usb, ethernet,...) with trait methods that are common between Connection
types.

Started on implementation of Connection for UsbConnection, some bugs need to
be resolved.

Started to switch Sever struct to expect object with Connection trait
rather than a specific type of connection to make it more general. Still needs
more work to be able to select between different connection types.

Separated functions specific to the haptic protocol out of UsbConnection and
into its own struct. Plan to have the Protocol have a connection field to
be able to send commands in protocol functions.