Continued with abstraction connection
Added protocol field to Server. Protocol has field for conn, which is
the same as the Server's to link the connection and protocol.
Protocol field will also need more generalized setup, but for now it expects
the only protocol implemented.
Calls to the the protocols methods were fixed in Server's serve() method.

Still has some errors. (lifetime, ownership, memory allocation).