Fixed ordering of bytes when building command
Created temporary vectors to hold the 4 bytes per memory block for the
three writable blocks. The memory blocks vectors are filled in order and
logic was added to fill the blocks depending on the type of command (
just timing, just actuators, both). Ex. Just timing uses 2 bytes of the
first memory block since command isn't used, where as an actuator command
only uses 1 byte since the command byte is used.

Memory blocks are reversed so the order is correct and added to the data
vector.

Mock was updated for testing purposes with a physical device.