Removed cache for actuator blocks
Using the cache for the actuator blocks caused errors/panic since previous
blocks need to be written, even when unchanged, to keep the order of the
commands. Using cache will result in None if not changed.
Extra blocks are truncated out after the last non-zero byte which reduces
the bytes sent without the cache.